crop-disease-backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.4.2
opencv-python==4.8.1.78
scikit-learn==1.3.2
python-magic==0.4.27
celery==5.3.4
redis==5.0.1
pydantic-settings==2.1.0